server/requirements.txt
django~=3.0.5
psycopg2-binary>=2.8
djangorestframework~=3.12.2
djongo~=1.3.4
djangorestframework-jwt~=1.11.0
random-object-id~=2.0.0